refactor: Update airflow dag to only download crime data if not exists

+++ src/nyc_crime_airflow_etl.py
@@ -32,17 +32,28 @@ def download_store_nyc_crime(**kwargs):
     None
     """
 
-    logging.info("Started download and storage of nyc crime data to S3 compatible storage backend")
-
-    # Download NYC crime data
-    urllib.request.urlretrieve("https://data.cityofnewyork.us/api/views/qgea-i56i/rows.csv?accessType=DOWNLOAD", "/tmp/nyc-crime-new-airflow.csv")
-
-    # Upload generated file to Minio
     # S3 Hook should called "local_minio" should be created already on Airflow
     s3 = S3Hook('local_minio')
-    s3.load_file("/tmp/nyc-crime-new-airflow.csv",
-                 key=f"nyc-crime-new-airflow.csv",
-                 bucket_name="nyc-crime")
+
+    # Check if bucket already exists
+    # If not, create new bucket
+    bucket_exists = s3.check_for_bucket("nyc-crime")
+    if not bucket_exists:
+        s3.create_bucket("nyc-crime")
+
+    file_exists = s3.check_for_key(f"nyc-crime-new-airflow.csv", bucket="nyc-crime")
+    if not file_exists:
+
+        logging.info("Started download and storage of nyc crime data to S3 compatible storage backend")
+
+        # Download NYC crime data
+        urllib.request.urlretrieve("https://data.cityofnewyork.us/api/views/qgea-i56i/rows.csv?accessType=DOWNLOAD", "/tmp/nyc-crime-new-airflow.csv")
+        s3.load_file("/tmp/nyc-crime-new-airflow.csv",
+                     key=f"nyc-crime-new-airflow.csv",
+                     bucket_name="nyc-crime")
+    else:
+
+        logging.info("File already exists")
 
 
 # DAG task to download to S3 compatible storage backend
